Split out webpack config into separate files.
