+++ package.json
@@ -13,7 +13,8 @@
     "react": "^15.0.0",
     "react-dom": "^15.0.0",
     "roneo": "icelab/roneo",
-    "viewloader": "^1.1.0"
+    "viewloader": "^1.1.0",
+    "webpack-merge": "0.12.0"
   },
   "devDependencies": {
     "babel-core": "^6.4.0",
@@ -36,7 +37,6 @@
     "gulp-rev-all": "^0.7.6",
     "gulp-rev-rails-manifest": "0.0.1",
     "gulp-tap": "^0.1.3",
-    "html-loader": "^0.4.3",
     "imports-loader": "^0.6.5",
     "js-yaml": "^1.0.2",
     "postcss": "^5.0.14",
@@ -47,6 +47,7 @@
     "run-sequence": "^1.0.2",
     "style-loader": "^0.13.0",
     "webpack": "^1.13.0",
+    "webpack-config": "4.0.0",
     "webpack-dev-server": "1.14.0",
     "webpack-notifier": "^1.2.1"
   },
